+++ compiler/parser/parser.v
@@ -10,13 +10,14 @@ import compiler.token
 
 pub struct Parser {
 mut:
-	sf       &ast.SourceFile = 0
-	scanner  Scanner
-	prev_tok token.Token
-	tok      token.Token
-	peek_tok token.Token
-	scope    &ast.Scope
-	labels   []string
+	sf            &ast.SourceFile = 0
+	scanner       Scanner
+	prev_tok      token.Token
+	tok           token.Token
+	peek_tok      token.Token
+	scope         &ast.Scope
+	labels        []string
+	inside_anon_t bool
 }
 
 pub fn run_parser() {
@@ -106,7 +107,7 @@ fn (mut p Parser) parse_symbol() &ast.Symbol {
 			p.next()
 		}
 		.mod {
-			if p.scope.is_root {
+			if p.scope.is_root && !p.inside_anon_t {
 				report.error('local scope prefix (`%`) cannot be used in global scope',
 					pos).emit()
 			}
@@ -272,6 +273,7 @@ fn (mut p Parser) parse_type() ast.Type {
 		return ast.Type(g_context.find_or_register_array(elem_typ, size))
 	} else if p.accept(.lbrace) {
 		// anonymous type: { i32 %field1, bool %field2 }
+		p.inside_anon_t = true
 		mut fields := []&ast.Symbol{}
 		if p.tok.kind != .rbrace {
 			for {
@@ -288,6 +290,7 @@ fn (mut p Parser) parse_type() ast.Type {
 			}
 		}
 		p.check(.rbrace)
+		p.inside_anon_t = false
 		return ast.Type(g_context.find_or_register_struct_type(ast.StructInfo{fields}))
 	}
 	prefix := p.tok.kind
@@ -489,12 +492,25 @@ fn (mut p Parser) parse_global_assign() ast.Stmt {
 	match instr {
 		'const' {
 			kind = .const_
-			expr = p.parse_literal()
 			left.kind = .constant
+			expr = p.parse_literal()
 		}
 		'type' {
 			kind = .type_
 			left.kind = .type_
+			mut tpos := p.tok.position()
+			typ := p.parse_type()
+			tpos = tpos.extend(p.prev_tok.position())
+			expr = ast.TypeNode{
+				typ: typ
+				pos: tpos
+			}
+			left.typ = ast.Type(g_context.register_type_symbol(ast.TypeSymbol{
+				name: left.name[1..]
+				gname: left.gname
+				kind: .alias
+				info: ast.AliasInfo{typ}
+			}))
 		}
 		else {
 			report.error('invalid instruction for global symbols', instr_pos).emit()