+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-* Merge the `goto` instruction with `br` instruction.
-* Added `getelement` instruction, to access elements of arrays (soon also for fields of structs).
+* **CHANGED:** Merge the `goto` instruction with the `br` instruction.
+* **ADDED:** `getelement` instruction, to access elements of arrays.
+* **ADDED:** Documentation (`docs/` folder).
 
 ## [0.1.0] - 2021-07-21