+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 * **CHANGED:** Merge the `goto` instruction with the `br` instruction.
 * **ADDED:** `getelement` instruction, to access elements of arrays.
 * **ADDED:** Documentation (`docs/` folder).
+* **ADDED:** Support for constants.
 
 ## [0.1.0] - 2021-07-21